DEX-663 respond to code review feedback and DRY up notification props
