+++ src/components/dialogs/NotificationDetailsDialog.jsx
@@ -9,6 +9,7 @@ import StandardDialog from '../StandardDialog';
 import Text from '../Text';
 import Button from '../Button';
 import CustomAlert from '../Alert';
+import { getNotificationProps } from '../../utils/notificationUtils';
 
 export default function NotificationDetailsDialog({
   open,
@@ -25,31 +26,13 @@ export default function NotificationDetailsDialog({
     onClose();
   };
 
-  const senderName = get(notification, 'sender_name', 'Unnamed User');
-  let user1Name = get(
-    notification,
-    ['message_values', 'user1_name'],
-    'Unnamed User',
-  );
-  if (user1Name === '') user1Name = 'Unnamed User';
-  let user2Name = get(
-    notification,
-    ['message_values', 'user2_name'],
-    'Unnamed User',
-  );
-  if (user2Name === '') user2Name = 'Unnamed User';
-  const individual1Names = notification?.names || []; // TODO flesh out more once this is included in notifications DEX-739
-  const individual1NicknameObject = individual1Names.find(
-    n => n.context === 'nickname',
-  );
-  const individual1Nickname =
-    individual1NicknameObject?.value || 'Unnamed individual';
-  const individual2Names = notification?.names || []; // TODO flesh out more once this is included in notifications DEX-739
-  const individual2NicknameObject = individual2Names.find(
-    n => n.context === 'nickname',
-  );
-  const individual2Nickname =
-    individual2NicknameObject?.value || 'Unnamed individual';
+  const {
+    senderName,
+    user1Name,
+    user2Name,
+    individual1Nickname,
+    individual2Nickname,
+  } = getNotificationProps(notification);
 
   return (
     <StandardDialog
