+++ src/pages/notifications/Notifications.jsx
@@ -24,6 +24,7 @@ import { calculatePrettyTimeElapsedSince } from '../../utils/formatters';
 import { notificationSchema } from '../../constants/notificationSchema';
 import { notificationTypes } from '../../components/dialogs/notificationDialogUtils';
 import queryKeys from '../../constants/queryKeys';
+import { getNotificationProps } from '../../utils/notificationUtils';
 
 export default function Notifications() {
   const intl = useIntl();
@@ -86,35 +87,13 @@ export default function Notifications() {
                   notificationType,
                 );
                 const read = get(notification, 'is_read', false);
-                const senderName = get(
-                  notification,
-                  'sender_name',
-                  'Unnamed User',
-                );
-                const user1Name = get(
-                  notification,
-                  ['message_values', 'user1_name'],
-                  'Unnamed User',
-                );
-                const user2Name = get(
-                  notification,
-                  ['message_values', 'user2_name'],
-                  'Unnamed User',
-                );
-                const individual1Names = notification?.names || []; // TODO flesh out more once this is included in notifications DEX-739
-                const individual1NicknameObject = individual1Names.find(
-                  n => n.context === 'nickname',
-                );
-                const individual1Nickname =
-                  individual1NicknameObject?.value ||
-                  'Unnamed individual';
-                const individual2Names = notification?.names || []; // TODO flesh out more once this is included in notifications DEX-739
-                const individual2NicknameObject = individual2Names.find(
-                  n => n.context === 'nickname',
-                );
-                const individual2Nickname =
-                  individual2NicknameObject?.value ||
-                  'Unnamed individual';
+                const {
+                  senderName,
+                  user1Name,
+                  user2Name,
+                  individual1Nickname,
+                  individual2Nickname,
+                } = getNotificationProps(notification);
                 const createdDate = notification?.created;
                 const timeSince = calculatePrettyTimeElapsedSince(
                   createdDate,
